add exception handler

+++ src/main/java/edu/examples/orders/repository/CustomerRepository.java
@@ -5,4 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.querydsl.QuerydslPredicateExecutor;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer>, JpaSpecificationExecutor<Customer>, QuerydslPredicateExecutor<Customer> {}
+import java.util.Optional;
+
+public interface CustomerRepository extends JpaRepository<Customer, Long>, JpaSpecificationExecutor<Customer>, QuerydslPredicateExecutor<Customer> {
+}
